more tests and some comments about the disabled stuff

+++ AT-Forum-functions/functions/src/roles/role.service.ts
@@ -1,12 +1,13 @@
 import {RoleRepository} from "./role.repository";
 
 export class RoleService {
-    constructor(private roleRepository: RoleRepository) {
-
-    }
-
+    constructor(private roleRepository: RoleRepository) {}
 
     deleteRole(userId: string): Promise<void> {
+        if (userId === '' || userId === undefined) {
+            const error = new TypeError('Id has to be defined');
+            return Promise.reject(error);
+        }
         return this.roleRepository.deleteRole(userId);
     }
 }
